Add tests for blacklisting API
