+++ src/lib/busStop.ts
@@ -1,0 +1,6 @@
+import type { Coordinates } from "./location";
+
+export type BusStop = {
+	coordinates: Coordinates;
+	times: Date[];
+}
